Fixing issue with datatables loading spinner
