+++ brew_view/static/src/js/configs/dt_renderer.js
@@ -1,0 +1,36 @@
+
+runDTRenderer.$inject = [
+  'DTRendererService',
+];
+
+/**
+ * runDTRenderer - Tweak datatables rendering
+ * @param  {Object} DTRendererService    Data-tables' rendering service.
+ */
+export default function runDTRenderer(DTRendererService) {
+  DTRendererService.registerPlugin({
+    postRender: function(options, result) {
+      // Insert a spinner thingy next to the search box
+      let spinner = $('<span>')
+        .attr('id', 'dtSpinner')
+        .addClass('fa fa-spinner fa-pulse')
+        .css('margin-right', '5px')
+        .css('visibility', 'hidden');
+
+      $('.dataTables_filter label').prepend(spinner);
+
+      // Register callback to show / hide spinner thingy
+      let processingDelay = null;
+      $('.dataTable').on('processing.dt', function(e, settings, processing) {
+        if (!processing) {
+          clearTimeout(processingDelay);
+          $('#dtSpinner').css('visibility', 'hidden');
+        } else {
+          processingDelay = setTimeout(function() {
+            $('#dtSpinner').css('visibility', 'visible');
+          }, 500);
+        }
+      });
+    },
+  });
+};
